LLM_FastAPI_App/requirements.txt
langchain-google-genai
python-dotenv 
langchain-core
langchain-community
pypdf
streamlit
langchain-ollama
unstructured[all-docs]
fastapi
uvicorn
python-multipart
passlib
sqlalchemy
bcrypt
python-jose